powered probe holder model files
jig for holding and powering emote from usb. 
take 3 = main carriage 
mated back plate = what it says on the tin
mounting hardware = piece that bonds to the usb A breakout board and attaches to underside of take 3

assembly notes: copper tape is placed on contact sides of nub on take 3 to deliver power to the probe. solder wires to them and attached to the power terminals on the usb breakout, running the wire through the square holes on either side of the nub.